Moved attr representation file and fixed issue with unknown units
