+++ AoE2ScenarioParser/helper/helper.py
@@ -1,5 +1,6 @@
 import math
 from enum import Enum
+from typing import Union
 
 from AoE2ScenarioParser.datasets.buildings import BuildingInfo
 from AoE2ScenarioParser.datasets.heroes import HeroInfo
@@ -26,7 +27,7 @@ def i_to_xy(i, map_size):
 =============================================================="""
 
 
-def get_enum_from_unit_const(const: int) -> InfoDatasetBase:
+def get_enum_from_unit_const(const: int) -> Union[InfoDatasetBase, None]:
     """
     Returns an Enum corresponding with the given Const.
 
@@ -39,6 +40,7 @@ def get_enum_from_unit_const(const: int) -> InfoDatasetBase:
             return enum.from_id(const)
         except (ValueError, KeyError):
             continue
+    return None
 
 
 def get_int_len(num):
