+++ AoE2ScenarioParser/settings.py
@@ -7,6 +7,8 @@ PRINT_STATUS_UPDATES = True
 # Warning related settings
 DISABLE_WARNINGS = False
 """Disable all warnings thrown by the printers.warn function."""
+DISABLE_VERSION_WARNINGS = False
+"""Disable warnings about python versions"""
 RAISE_ERROR_ON_WARNING = False
 """Raise an error when a warning should be shown. Handy for debugging, shows you the stacktrace."""
